[auxia proxy] code layout refactoring
